Exclude patch versions on dependencies to reduce renovate spam
Since this repository neither contains private code nor do our actions
run with secrets, it is low risk to not have to manually approve a
renovate PR for every patch version. I also switched to an official
dprint action instead of using the mise action, as renovate was
not seeing that custom in-line config when updating and I don't see a
particular reason not to use the official action instead.